[CAS-245] Deleting AttachmentsProgressViewHolder
